@@ -1,7 +1,7 @@
 {
   "name": "spacemesh_app",
   "productName": "Spacemesh",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "author": {
     "name": "Spacemesh",
     "email": "info@spacemesh.io",
@@ -50,32 +50,31 @@
   ],
   "homepage": "https://spacemesh.io/",
   "devDependencies": {
-    "@babel/core": "7.9.6",
-    "@babel/plugin-proposal-class-properties": "7.8.3",
-    "@babel/plugin-proposal-export-namespace-from": "7.8.3",
-    "@babel/plugin-proposal-optional-chaining": "7.9.0",
-    "@babel/plugin-syntax-import-meta": "7.8.3",
-    "@babel/plugin-transform-react-constant-elements": "7.8.3",
-    "@babel/plugin-transform-react-inline-elements": "7.8.3",
-    "@babel/preset-env": "7.9.6",
-    "@babel/preset-flow": "7.9.0",
-    "@babel/preset-react": "7.9.4",
-    "@babel/register": "7.8.3",
-    "@hot-loader/react-dom": "16.13.0",
+    "@babel/core": "7.12.3",
+    "@babel/plugin-proposal-class-properties": "7.12.1",
+    "@babel/plugin-proposal-export-namespace-from": "7.12.1",
+    "@babel/plugin-proposal-optional-chaining": "7.12.1",
+    "@babel/plugin-syntax-import-meta": "7.10.4",
+    "@babel/plugin-transform-react-constant-elements": "7.12.1",
+    "@babel/plugin-transform-react-inline-elements": "7.12.1",
+    "@babel/preset-env": "7.12.1",
+    "@babel/preset-flow": "7.12.1",
+    "@babel/preset-react": "7.12.1",
+    "@babel/register": "7.12.1",
+    "@hot-loader/react-dom": "17.0.0",
     "aes-js": "3.1.2",
     "babel-eslint": "10.1.0",
     "babel-loader": "8.1.0",
-    "babel-plugin-styled-components": "1.10.7",
+    "babel-plugin-styled-components": "1.11.1",
     "babel-root-slash-import": "1.1.0",
     "cross-env": "7.0.2",
     "detect-port": "1.3.0",
     "dotenv": "8.2.0",
-    "electron": "8.5.2",
-    "electron-builder": "22.7.0",
-    "electron-devtools-installer": "3.1.0",
-    "electron-notarize": "0.3.0",
+    "electron": "10.1.5",
+    "electron-builder": "22.9.1",
+    "electron-devtools-installer": "3.1.1",
+    "electron-notarize": "1.0.0",
     "electron-unhandled": "3.0.2",
-    "electron-updater": "4.3.1",
     "eslint": "7.4.0",
     "eslint-config-airbnb": "18.2.0",
     "eslint-config-prettier": "6.11.0",
@@ -88,17 +87,17 @@
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-react": "7.20.3",
     "eslint-plugin-react-hooks": "4.0.7",
-    "file-loader": "6.0.0",
-    "flow-bin": "0.129.0",
+    "file-loader": "6.1.1",
+    "flow-bin": "0.136.0",
     "flow-runtime": "0.17.0",
-    "flow-typed": "3.2.0",
+    "flow-typed": "3.2.1",
     "hoist-non-react-statics": "3.3.2",
-    "prettier": "2.0.5",
-    "react-hot-loader": "4.12.21",
+    "prettier": "2.1.2",
+    "react-hot-loader": "4.13.0",
     "terser-webpack-plugin": "3.0.6",
-    "url-loader": "4.1.0",
+    "url-loader": "4.1.1",
     "webpack": "4.43.0",
-    "webpack-bundle-analyzer": "3.8.0",
+    "webpack-bundle-analyzer": "3.9.0",
     "webpack-cli": "3.3.12",
     "webpack-dev-server": "3.11.0",
     "webpack-merge": "4.2.2"
@@ -108,20 +107,21 @@
     "auto-launch": "5.0.5",
     "bip39": "3.0.2",
     "check-disk-space": "2.1.0",
-    "electron-log": "4.2.1",
-    "electron-store": "5.1.1",
-    "find-process": "1.4.3",
-    "grpc": "1.24.3",
+    "electron-log": "4.2.4",
+    "electron-store": "6.0.1",
+    "find-process": "1.4.4",
+    "@grpc/grpc-js": "1.1.7",
     "js-xdr": "1.1.4",
     "pbkdf2": "3.1.1",
-    "react": "16.8.6",
+    "react": "16.14.0",
     "react-beautiful-dnd": "11.0.5",
-    "react-dom": "16.8.6",
-    "react-redux": "7.2.0",
+    "react-dom": "16.14.0",
+    "react-redux": "7.2.1",
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
     "redux": "4.0.5",
     "redux-thunk": "2.3.0",
+    "regenerator-runtime": "^0.13.7",
     "styled-components": "5.1.1"
   }
 }
