@@ -1,5 +1,5 @@
 // @flow
-import React, { PureComponent } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import { Banner } from '/basicComponents';
 import { smColors } from '/vars';
@@ -12,16 +12,10 @@ const Text = styled.div`
   text-align: center;
 `;
 
-type Props = {};
-
-class LoggedOutBanner extends PureComponent<Props> {
-  render() {
-    return (
-      <Banner margin={'0 0 30px 0'} color={smColors.blue}>
-        <Text>YOU ARE LOGGED OUT FROM YOUR WALLET</Text>
-      </Banner>
-    );
-  }
-}
+const LoggedOutBanner = () => (
+  <Banner margin={'0 0 30px 0'} color={smColors.blue}>
+    <Text>YOU ARE LOGGED OUT FROM YOUR WALLET</Text>
+  </Banner>
+);
 
 export default LoggedOutBanner;
