@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React from 'react';
 import { SecondaryButton } from '/basicComponents';
 import { chevronLeftWhite } from '/assets/images';
 
@@ -10,16 +10,6 @@ type Props = {
 
 const style = { position: 'absolute', bottom: 0, left: -45 };
 
-class BackButton extends PureComponent<Props> {
-  static defaultProps = {
-    width: 10,
-    height: 15
-  };
-
-  render() {
-    const { action, width, height } = this.props;
-    return <SecondaryButton onClick={action} img={chevronLeftWhite} imgWidth={width} imgHeight={height} style={style} />;
-  }
-}
+const BackButton = ({ action, width = 10, height = 15 }: Props) => <SecondaryButton onClick={action} img={chevronLeftWhite} imgWidth={width} imgHeight={height} style={style} />;
 
 export default BackButton;
