@@ -72,7 +72,7 @@ class VaultMasterAccounts extends Component<Props, State> {
       <>
         <DetailsRow>
           <DetailsText>Address 1</DetailsText>
-          <Tooltip top="-2" left="-3" width="250" text="Use an account managed by this wallet to set yourself as the vault’s owner." />
+          <Tooltip width="250" text="Use an account managed by this wallet to set yourself as the vault’s owner." />
           <Dots />
           <DropDown
             data={masterAccounts}
@@ -86,7 +86,7 @@ class VaultMasterAccounts extends Component<Props, State> {
         </DetailsRow>
         <DetailsRow>
           <DetailsText>Address 2</DetailsText>
-          <Tooltip top="-2" left="-3" width="250" text="Use an account managed by this wallet to set yourself as the vault’s owner." />
+          <Tooltip width="250" text="Use an account managed by this wallet to set yourself as the vault’s owner." />
           <Dots />
           <DropDown
             data={masterAccounts}
@@ -100,7 +100,7 @@ class VaultMasterAccounts extends Component<Props, State> {
         </DetailsRow>
         <DetailsRow>
           <DetailsText>Address 3</DetailsText>
-          <Tooltip top="-2" left="-3" width="250" text="Use an account managed by this wallet to set yourself as the vault’s owner." />
+          <Tooltip width="250" text="Use an account managed by this wallet to set yourself as the vault’s owner." />
           <Dots />
           <DropDown
             data={masterAccounts}
