@@ -2,7 +2,7 @@
 
 const developmentEnvironments = ['development', 'test'];
 
-const developmentPlugins = [require('react-hot-loader/babel'), require('babel-plugin-styled-components')];
+const developmentPlugins = [require('babel-plugin-styled-components'), require('react-hot-loader/babel')];
 
 const productionPlugins = [
   // babel-preset-react-optimize
@@ -17,12 +17,7 @@ module.exports = (api) => {
 
   return {
     presets: [
-      [
-        require('@babel/preset-env'),
-        {
-          targets: { electron: require('electron/package.json').version }
-        }
-      ],
+      require('@babel/preset-env'),
       require('@babel/preset-flow'),
       [require('@babel/preset-react'), { development }]
     ],
